screen-orientation/onchange-event.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(t => {
  const type = screen.orientation.type;
  // wrap onchange listener in new promise to catch unreached assertion
  return new Promise(async (resolve, reject) => {
    screen.orientation.onchange = () => { 
      reject(new Error("change event should not be fired"));
    }
    await screen.orientation.lock(type);
    assert_equals(screen.orientation.type, type);
    setTimeout(resolve, 1000);
  });
}, "Test that orientationchange event is not fired when the orientation does not change.");

promise_test(async t => {
  let orientations = [
    'portrait-primary',
    'portrait-secondary',
    'landscape-primary',
    'landscape-secondary'
  ];
  if (screen.orientation.type.includes('portrait')) {
    orientations = orientations.reverse();
  }
  const orientationWatcher = new EventWatcher(t, screen.orientation, 'change');

  for (const orientation of orientations) {
    screen.orientation.lock(orientation);
    await orientationWatcher.wait_for('change');
    assert_equals(screen.orientation.type, orientation);
  }
  screen.orientation.unlock();
}, "Test that orientationchange event is fired when the orientation changes.");
</script>
